microservices/auth_service/requirements.txt
# Auth Service Requirements
# UK Management Bot - Auth Service

# FastAPI and ASGI
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Redis for session storage
redis[hiredis]==5.0.1

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0  # TOTP/HOTP for multi-factor authentication

# HTTP Client for service communication
httpx==0.25.2

# Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # For testing HTTP requests
factory-boy==3.3.0  # For test data factories
black==23.11.0
ruff==0.1.6

# Logging
structlog==23.2.0

# Additional utilities
click==8.1.7